Update period-closing-voucher.md
Capital Fund account is a misnomer for Revenue reserve group. Profit & Loss balance is in the nature of Revenue reserve and it can not be merged into Capital reserve account. Another word for the reserve is also "Fund". Hence making these changes would make it accurate in the accounting sense.